+++ servsrc/server.py
@@ -3,19 +3,13 @@ import socketserver
 import logging
 import json
 import sqlite3
+from database import Dbase
+import cgi
 
-def gen_path(paths):
-    full_path = os.path.dirname(os.path.abspath(__file__))
-    for path in paths:
-        full_path = os.path.join(full_path, path)
-    return full_path
-
-
-class Serv(BaseHTTPRequestHandler):
+class Serv(http.server.BaseHTTPRequestHandler):
     def __init__(self, *args, **kwargs):
         super(Serv, self).__init__(*args, **kwargs)
-        self.log_path = gen_path(['logs', 'serv_work.log'])
-        self.users_db = sqlite3.connect(gen_path(['..', 'data', 'users.db'])
+        #self.dbase = Dbase()
 
     def _set_headers(self):
         self.send_response(200)
@@ -26,14 +20,14 @@ class Serv(BaseHTTPRequestHandler):
         self._set_headers()
 
     def do_POST(self):
-        pass
+        length = int(self.headers.get('content-length'))
+        message = json.loads(self.rfile.read(length))
+        message['received'] = 'ok'
+        self._set_headers()
+        self.wfile.write(json.dumps(message).encode('utf-8'))
 
-    def create_users_db(self):
 
-
-
-
-def run(server_class = http.server.HTTPServer, handler_class = http.server.BaseHTTPRequestHandler, port = 8000):
+def run(log_path, server_class = http.server.HTTPServer, handler_class = Serv, port = 8000):
     logging.basicConfig(filename = log_path, filemode = 'w', level = logging.DEBUG)
     httpd = server_class(('localhost', port), handler_class)
     try:
